use Object.literal instead of Object.create(null)
The resaon behind this change is that I'm using for sorting JSON objects,
and comparing them with `assert.deepStrictEqual` method (introduced in v4)
and I expected tests to pass. But as far current implmentation
creates objects without prototype it doesn't work this way.
And it seems wrong to me.

So I wanted to start discussion about both approaches.